fix some bugs

+++ install.sh
@@ -2,6 +2,7 @@
 
 echo "start Installing......"
 
+
 # install git vim zsh curl
 if which yum >/dev/null; then
     sudo yum install -y git vim zsh curl >/dev/null
@@ -9,29 +10,39 @@ elif which apt-get >/dev/null; then
     sudo apt-get install -y git vim zsh curl >/dev/null
 fi
 
+
 if [ -d "~/vim" ]; then
-    mv -f ~/vim ~/vim_old
+    mv ~/vim ~/vim_old
 fi
-cd ~/ && git clone https://github.com/if-tao/vim.git
-mv -f ~/.vim ~/.vim_old
+git clone https://github.com/if-tao/vim.git ~/vim
+if [ -d "~/.vim" ]; then
+    tar -zcPf ~/.vim_old.tar.gz ~/.vim
+fi
+if [ -e "~/.vimrc" ]; then
+    mv -f ~/.vimrc ~/.vimrc_old
+fi
 mv -f ~/vim ~/.vim
-mv -f ~/.vimrc ~/.vimrc_old
-mv -f ~/.vim/.vimrc ~/
+cp -f ~/.vim/.vimrc ~/.vimrc
+
+
 git clone https://github.com/VundleVim/Vundle.vim.git ~/.vim/bundle/Vundle.vim
 echo "正在配置vim插件..." > if-tao
 echo "请耐心等待..." >> if-tao
 vim if-tao -c "PluginInstall" -c "q" -c "q"
 rm if-tao
 
-cd ~/.vim/
-if [ -d "~/.vim/vimcdoc/vimcdoc-1.9.0/" ]; then
-    mkdir ~/.vim/vimcdoc/vimcdoc-1.9.0
-tar -zxPf ~/.vim/vimcdoc/vimcdoc-1.9.0.tar.gz ~/.vim/vimcdoc/vimcdoc-1.9.0/
-bash ~/.vim/vimcdoc/vimcdoc-1.9.0/vimcdoc.sh
+
+cd ~/.vim/vimcdoc/
+tar -zxPf ./vimcdoc-1.9.0.tar.gz
+cd ./vimcdoc-1.9.0/
+bash ./vimcdoc.sh -i >>/dev/null
 rm -rf ~/.vim/vimcdoc/vimcdoc-1.9.0/
+cd ~/
 
 # oh_my_zsh
 echo "正在配置oh_my_zsh..."
 sh -c "$(curl -fsSL https://raw.github.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 
+
 echo "Install success !!"
+
